requirements.txt
flask==3.0.0
bootstrap_flask==2.3.2
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
numpy==1.26.2
pillow==10.1.0

